Move `activecate` -> `activegate/coreReconciler`
