+++ src/controllers/dynakube/activegate/coreReconciler/reconciler.go
@@ -1,4 +1,4 @@
-package activegate
+package coreReconciler
 
 import (
 	"context"
